Feat: Added method to fetch shipment details by token
Added a new method named getShipmentByToken to
the ShipmentsController.php file. This method
fetches shipment details and related data from
the database based on the provided token. If
the shipment is found, it returns the details
as a JSON response. If the shipment is not
found, it returns an error message.

Changes made:
- Added a new method getShipmentByToken to
  fetch shipment details by token.