Add CitySeeder to Database Seeder
Adds the CitySeeder class to the DatabaseSeeder
file. By including the CitySeeder in the database
seeder, we ensure that cities are seeded along with
other essential data when running the database seeder.
This helps maintain consistency in the seeding process
and ensures that all necessary data is populated in the
database during development and testing.