fix: Update About Us link in Welcome.vue to point to the correct page
This commit fixes the About Us link in the
Welcome.vue component to point to the correct page.
Previously, the link was pointing to the register
page, as the About Us page was not yet created.
With this commit, the About Us page has been
created, and the link in Welcome.vue has been
updated accordingly to route users to the
About Us page.