feat: Add route for About page in web.php
This commit adds a new route in the web.php file to
handle requests for the About page. The route
directs to the AboutController@index method,
allowing users to access the About page by visiting
the '/About' URL. This addition enables users to
navigate to the About page and learn more about the
application.