Feat: Update Contact Us link in footer.vue to point to the correct route
This commit updates the Contact Us link in the
footer.vue component to point to the correct
route. Previously, the link was pointing to "#"
as a placeholder, but now it has been updated
to route users to the "/Contact" page, which is
the route configured in the application for the
Contact Us page.