Add email notifications for shipment status updates
Modified the ShipmentsController to include email
notifications for various shipment status updates.
Imported the necessary mail classes for sending
notifications when a shipment is out for delivery
to both sender and recipient, as well as when
a shipment is delivered. Updated the updateStatus
method to send these notifications based on the
status of the shipment.