Add RecipientShipmentOutForDelivery mail class
Created a new mail class,
RecipientShipmentOutForDelivery, responsible for
sending email notifications to the recipient when
before a shipment is out for delivery. This class
accepts a Shipment instance in its constructor to
provide the necessary data for composing the email.
The build method specifies the markdown template to
se for the email and sets the subject to "Shipment Out
for Delivery".