Add SenderShipmentOutForDelivery mail class
Introduce a new mail class,
SenderShipmentOutForDelivery, designed to send
email notifications to the sender when a shipment
is out for delivery. The constructor accepts
a Shipment instance to provide the necessary data
for composing the email. The build method specifies
the markdown template to use for the email and sets
the subject to "Shipment Out for Delivery".