Add ShipmentDelivered mail class
Introduce a new mail class, ShipmentDelivered,
responsible for sending email notifications when
a shipment has been successfully delivered to the
recipient. The constructor accepts a Shipment
instance and the recipient's name to provide the
necessary data for composing the email. The build
method specifies the markdown template to use for
the email and sets the subject to "Shipment Delivered".